requirements.txt
requests
beautifulsoup4
pandas
openai
python-dateutil
chromadb
streamlit
tiktoken
python-dotenv
